Create notification && subscribe With test and Fix bugs

+++ src/app/Http/Controllers/API/v1/Thread/AnswerController.php
@@ -4,14 +4,27 @@ namespace App\Http\Controllers\API\v1\Thread;
 
 use App\Http\Controllers\Controller;
 use App\Models\Answer;
+use App\Models\Subscribe;
 use App\Models\Thread;
+use App\Models\User;
+use App\Notifications\NewReplySubmitted;
 use App\Repositories\AnswerRepository;
+use App\Repositories\SubscribeRepository;
+use App\Repositories\UserRepository;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Notification;
 
 class AnswerController extends Controller
 {
+
+    public function __construct(){
+        $this->middleware(['user_block'])->except([
+            'index'
+        ]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -37,6 +50,19 @@ class AnswerController extends Controller
             'thread_id'=>'required',
         ]);
         resolve(AnswerRepository::class)->store($request);
+        //Get List of User Id Which Subscribed of A Thread Id
+        $notifible_users_id=resolve(SubscribeRepository::class)->getNotifibleUsers($request->thread_id);
+        //Get User Instace From Id
+        $notifible_users=resolve(UserRepository::class)->find($notifible_users_id);
+        //Send NewReplySubmitted Notifi Subscribed Users
+        Notification::send($notifible_users,new NewReplySubmitted(Thread::find($request->thread_id)));
+
+        //Increase User Score
+        if(Thread::find($request->input('thread_id'))->user_id!== auth()->id()){
+            auth()->user()->increment('score',10);
+        }
+
+
         return response()->json([
             'message'=>'answer submitted successfully'
         ], Response::HTTP_CREATED);
@@ -55,11 +81,14 @@ class AnswerController extends Controller
             'content'=>'required',
         ]);
 
-        resolve(AnswerRepository::class)->update($request,$answer);
+        if(Gate::forUser(auth()->user())->allows('user-answer',$answer)) {
+            resolve(AnswerRepository::class)->update($request, $answer);
 
-        return response()->json([
-            'message'=>'answer updated successfully'
-        ], Response::HTTP_OK);
+            return response()->json([
+                'message' => 'answer updated successfully'
+            ], Response::HTTP_OK);
+        }
+        return response()->json(['massage'=>'Access Denied',], Response::HTTP_FORBIDDEN);
     }
 
     /**
@@ -70,9 +99,12 @@ class AnswerController extends Controller
      */
     public function destroy(Answer $answer)
     {
+        if(Gate::forUser(auth()->user())->allows('user-answer',$answer)) {
         resolve(AnswerRepository::class)->destroy($answer);
         return response()->json([
             'message'=>'answer deleted successfully'
         ], Response::HTTP_OK);
+        }
+        return response()->json(['massage'=>'Access Denied',], Response::HTTP_FORBIDDEN);
     }
 }
